requirement.txt
djangorestframework
markdown
django-filter
daphne
channels
django-cors-headers
djangorestframework-simplejwt
sentence_transformers
datasets
transformers[torch]
